+++ changelog.md
@@ -3,6 +3,10 @@
 ## version 1.1.x
 notes from version 1.1.0 +
 
+### version 1.1.4
+- add $audio in core module & browser
+- add webcomponent getJson > warning if api using cors
+
 ### version 1.1.3
 - add $url in core module, node & browser
 - add $browserChecker in browser & core